requirements.txt
asgiref==3.3.4
Django==3.2.2
pytz==2021.1
sqlparse==0.4.1
transbank-sdk==1.4.0
Pillow==8.2.0
docker==5.0.0
docker-compose==1.29.2